break out implied-superview shortcuts
